Begin creating a C++ wrapper.
This is necessary because a lot of the functionality we need bindgen (or
Rust itself) cannot provide. Most prominently, constructing wxWidgets'
objects from Rust is very difficult. We can't call the constructors
themselves from Rust (constructors don't use the normal calling
convention), and the objects make heavy use of inheritance, which we
really can't handle.

It's also worth mentioning that we need C++ to handle most allocation
for us, since many wxWidgets objects cannot be moved, and can be freed
from C++ when e.g. their container is destroyed.

Currently, much of this wrapper is generated from Doxygen XML output.
This approach seems dubious, but it's what wxPython does, so I guess
it's fine. It does suffer somewhat when the docs are wrong, but at least
it only binds to intentionally public interfaces.